Merging Documentation Changes, Refactoring and other general Code Changes

+++ trim-deploy.sh
@@ -1,5 +1,6 @@
 while IFS="" read -r file || [ -n "$file" ]
 do
+  [[ $file =~ ^#.* ]] && continue
   rm -rf $file
-  echo "$file Ignored from Deployment."
+  echo "$file Ignored from Deployment"
 done < .deployignore
